Add intention to replace snake-case test function name with a space-separated
#KT-27143 Fixed
