+++ idea/resources/META-INF/plugin-common.xml
@@ -1373,6 +1373,11 @@
     </intentionAction>
 
     <intentionAction>
+      <className>org.jetbrains.kotlin.idea.intentions.ConvertSnakeCaseTestFunctionToSpacedIntention</className>
+      <category>Kotlin</category>
+    </intentionAction>
+
+    <intentionAction>
       <className>org.jetbrains.kotlin.idea.intentions.ReplaceSingleLineLetIntention</className>
       <category>Kotlin</category>
     </intentionAction>
