+++ idea/resources/intentionDescriptions/ConvertSnakeCaseTestFunctionToSpacedIntention/after.kt.template
@@ -1,0 +1,5 @@
+class MyTest {
+    @Test fun <spot>`test two plus two equals four`</spot>() {
+
+    }
+}
