+++ idea/resources/intentionDescriptions/ConvertSnakeCaseTestFunctionToSpacedIntention/before.kt.template
@@ -1,0 +1,5 @@
+class MyTest {
+    @Test fun <spot>test_two_plus_two_equals_four</spot>() {
+
+    }
+}
