+++ idea/resources/intentionDescriptions/ConvertSnakeCaseTestFunctionToSpacedIntention/description.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+This intention changes a snake case name of the given test function to a space-separated one.
+</body>
+</html>
