+++ idea/testData/intentions/convertCamelCaseTestFunctionToSpaced/snake.kt
@@ -1,0 +1,12 @@
+// CONFIGURE_LIBRARY: JUnit@lib/junit-4.12.jar
+// IS_APPLICABLE: false
+
+import org.junit.Test
+
+class A {
+    @Test fun <caret>test_two_plus_two_equals_four() {}
+}
+
+fun test() {
+    A().test_two_plus_two_equals_four()
+}
