+++ idea/testData/intentions/convertSnakeCaseTestFunctionToSpaced/.intention
@@ -1,0 +1,1 @@
+org.jetbrains.kotlin.idea.intentions.ConvertSnakeCaseTestFunctionToSpacedIntention
