+++ idea/testData/intentions/convertSnakeCaseTestFunctionToSpaced/camel.kt
@@ -1,0 +1,8 @@
+// CONFIGURE_LIBRARY: JUnit@lib/junit-4.12.jar
+// IS_APPLICABLE: false
+
+import org.junit.Test
+
+class A {
+    @Test fun <caret>testTwoPlusTwoEqualsFour() {}
+}
